pyproject.toml
[project]
name = "policy-pilot-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "ruff>=0.15.4",
    "uvicorn>=0.41.0",
]
